+++ flexmeasures/cli/data_show.py
@@ -246,8 +246,6 @@ def show_generic_asset(asset):
         (
             asset.generic_asset_type.name,
             asset.location,
-            "".join([f"{k}: {v}\n" for k, v in asset.flex_context.items()]),
-            "".join([f"{k}: {v}\n" for k, v in asset.flex_model.items()]),
             "".join(
                 [
                     f"{graph['title']}: {graph['sensors']} \n"
@@ -263,10 +261,24 @@ def show_generic_asset(asset):
             headers=[
                 "Type",
                 "Location",
-                "Flex-Context",
-                "Flex-Model",
                 "Sensors to show",
                 "Attributes",
+            ],
+        )
+    )
+
+    flex_config = [
+        (
+            "".join([f"{k}: {v}\n" for k, v in asset.flex_context.items()]),
+            "".join([f"{k}: {v}\n" for k, v in asset.flex_model.items()]),
+        )
+    ]
+    click.echo(
+        tabulate(
+            flex_config,
+            headers=[
+                "Flex-Context",
+                "Flex-Model",
             ],
         )
     )