+++ documentation/cli/change_log.rst
@@ -11,7 +11,9 @@ since v0.30.0 | November xx, 2025
 * Fix the interpretation of ``--start-offset`` and ``--end-offset`` options in ``flexmeasures add report``, which was not using the correct time at which the command was run; instead, that time was off by and offset corresponding to the offset between the ``FLEXMEASURES_TIMEZONE`` config setting and :abbr:`UTC (Coordinated Universal Time)`).
 * Add `--force` option to ``flexmeasures delete beliefs`` 
 * The ``flexmeasures delete beliefs``, ``flexmeasures delete nan-beliefs`` and ``flexmeasures delete unchanged-beliefs`` CLI commands now support limiting the action to a given set of data sources.
+* Allow asset-level scheduling in ``flexmeasures add schedule`` using the ``--asset`` option.
 * Allow modifying asset trees with the CLI using ``flexmeasures edit transfer-parenthood``.
+* ``flexmeasures show asset`` will now also display the flex model. 
 
 since v0.29.0 | October 14, 2025
 =================================